+++ src/app.controller.spec.ts
@@ -1,0 +1,29 @@
+import { Logger } from '@nestjs/common'
+import { Test, TestingModule } from '@nestjs/testing'
+import { AppController } from './app.controller'
+import { AppService } from './app.service'
+
+describe('AppController', () => {
+    let appController: AppController
+
+    beforeEach(async () => {
+        const app: TestingModule = await Test.createTestingModule({
+            controllers: [AppController],
+            providers: [
+                AppService,
+                {
+                    provide: Logger,
+                    useValue: { log: jest.fn() }
+                }
+            ]
+        }).compile()
+
+        appController = app.get<AppController>(AppController)
+    })
+
+    describe('root', () => {
+        it('should return "Hello World!"', () => {
+            expect(appController.getHello()).toBe('Hello World!')
+        })
+    })
+})