+++ src/app.module.ts
@@ -1,17 +1,15 @@
-import { Module } from '@nestjs/common'
+import { Logger, Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { NacosModule } from '@wlisfes/chat-web-base-schema/nacos'
-import { HealthModule } from '@/modules/health/health.module'
-import { SkylineModule } from '@/modules/skyline/skyline.module'
-import { SsrModule } from '@/modules/ssr/ssr.module'
+import { AppController } from './app.controller'
+import { AppService } from './app.service'
 
 @Module({
     imports: [
         ConfigModule.forRoot({ isGlobal: true }),
-        NacosModule.forRoot({ serviceName: 'chat-web-skyline-service', registerPort: 4020 }),
-        SsrModule,
-        HealthModule,
-        SkylineModule
-    ]
+        NacosModule.forRoot({ serviceName: 'chat-web-skyline-service', registerPort: 3000 })
+    ],
+    controllers: [AppController],
+    providers: [Logger, AppService]
 })
 export class AppModule {}