+++ src/app.service.ts
@@ -1,0 +1,11 @@
+import { Injectable, Logger } from '@nestjs/common'
+
+@Injectable()
+export class AppService {
+    constructor(private readonly logger: Logger) {}
+
+    getHello(): string {
+        this.logger.log('正在获取欢迎信息', AppService.name)
+        return 'Hello World!'
+    }
+}